Updates - see comments
removed many unimplemented methods and added multiple pages for game functions.

Modified navigation.html to include the changes to html structure.

Modified test.html to include bootstrap classes

Updated the error page to be prettier and provide feedback to the user

Changed savegamedata to /savegamedata to provide accurate endpoint mapping (GET implied) added logging to show when data is being written to the database.

Need to focus on implementing connections between the service layer and user interface.

Need to remove hard-coded database writes and tie into implementations to be written in the service layer